# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OfferLoots GPT Rewards Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedUsers"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user detail (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin) - ban/unban, edit balance/points
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/offers:
    get:
      operationId: getAdminOffers
      tags: [admin]
      summary: List all offers (admin)
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAdminOffers"
    post:
      operationId: createAdminOffer
      tags: [admin]
      summary: Create offer (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /admin/offers/{id}:
    patch:
      operationId: updateAdminOffer
      tags: [admin]
      summary: Update offer (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteAdminOffer
      tags: [admin]
      summary: Delete offer (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/withdrawals:
    get:
      operationId: getAdminWithdrawals
      tags: [admin]
      summary: List all withdrawals (admin)
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedWithdrawals"
  /admin/withdrawals/{id}:
    patch:
      operationId: updateAdminWithdrawal
      tags: [admin]
      summary: Approve/reject withdrawal (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalStatusUpdate"
      responses:
        "200":
          description: Updated withdrawal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Withdrawal"
  /admin/conversions:
    get:
      operationId: getAdminConversions
      tags: [admin]
      summary: List all conversions (admin)
      parameters:
        - in: query
          name: network
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Conversions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAdminConversions"
...
  /admin/revenue:
    get:
      operationId: getAdminRevenue
      tags: [admin]
      summary: Revenue chart data by period
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [7d, 30d, 90d]
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenuePoint"
  /admin/networks:
    get:
      operationId: getAdminNetworks
      tags: [admin]
      summary: Get all offerwall networks (admin)
      responses:
        "200":
          description: Networks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Network"
    post:
      operationId: createAdminNetwork
      tags: [admin]
      summary: Create/configure a network (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NetworkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
  /admin/networks/{id}:
    patch:
      operationId: updateAdminNetwork
      tags: [admin]
      summary: Update network config (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NetworkUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
components:
  schemas:
...
    Offer:
      type: object
      required: [id, name, payout, network, status, category, device, countries, description, imageUrl]
      properties:
        id:
          type: integer
        name:
          type: string
        payout:
          type: number
        network:
          type: string
        networkId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [active, paused, deleted]
        category:
          type: string
        device:
          type: string
          enum: [all, mobile, desktop]
        countries:
          type: array
          items:
            type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        offerUrl:
          type: ["string", "null"]
        completions:
          type: integer
        createdAt:
          type: string
    OfferInput:
      type: object
      required: [name, payout, network, category, device, countries, description]
      properties:
        name:
          type: string
        payout:
          type: number
        network:
          type: string
        networkId:
          type: ["integer", "null"]
        category:
          type: string
        device:
          type: string
        countries:
          type: array
          items:
            type: string
        description:
          type: string
        imageUrl:
          type: string
        offerUrl:
          type: string
    OfferUpdate:
      type: object
      properties:
        name:
          type: string
        payout:
          type: number
        status:
          type: string
        category:
          type: string
        device:
          type: string
        countries:
          type: array
          items:
            type: string
        description:
          type: string
        imageUrl:
          type: string
        offerUrl:
          type: string
...
    PaginatedAdminOffers:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Conversions
...
    PaginatedAdminConversions:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminConversion"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminConversion:
      type: object
      required: [id, userId, username, offerId, offerName, network, amount, status, txid, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        offerId:
          type: ["integer", "null"]
        offerName:
          type: string
        network:
          type: string
        amount:
          type: number
        status:
          type: string
        txid:
          type: string
        ip:
          type: ["string", "null"]
        createdAt:
          type: string

    # Transactions
...
    Withdrawal:
      type: object
      required: [id, userId, amount, method, accountInfo, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [paypal, payoneer, wise, crypto_btc, crypto_eth, crypto_usdt, bkash, nagad, manual]
        accountInfo:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected, processing]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    WithdrawalStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected, processing]
        notes:
          type: string
    PaginatedWithdrawals:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Withdrawal"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # API Keys
...
    Network:
      type: object
      required: [id, name, slug, isActive, postbackUrl, totalConversions, totalRevenue]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        postbackUrl:
          type: string
        secretKey:
          type: ["string", "null"]
        totalConversions:
          type: integer
        totalRevenue:
          type: number
    NetworkInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
        secretKey:
          type: string
        isActive:
          type: boolean
    NetworkUpdate:
      type: object
      properties:
        name:
          type: string
        logoUrl:
          type: string
        secretKey:
          type: string
        isActive:
          type: boolean

    # Admin
...
    PaginatedUsers:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AdminUserSummary:
      type: object
      required: [id, username, email, status, role, balance, points, rank, totalEarned, totalConversions, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        status:
          type: string
        role:
          type: string
        balance:
          type: number
        points:
          type: integer
        rank:
          type: string
        totalEarned:
          type: number
        totalConversions:
          type: integer
        createdAt:
          type: string
    AdminUserDetail:
      type: object
      required: [id, username, email, status, role, balance, points, rank, totalEarned, totalWithdrawn, totalConversions, referralCode, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        status:
          type: string
        role:
          type: string
        balance:
          type: number
        points:
          type: integer
        rank:
          type: string
        totalEarned:
          type: number
        totalWithdrawn:
          type: number
        totalConversions:
          type: integer
        referralCode:
          type: string
        avatar:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        lastIp:
          type: ["string", "null"]
        createdAt:
          type: string
    AdminUserUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, banned]
        balance:
          type: number
        points:
          type: integer
        role:
          type: string
          enum: [user, admin]
...
    RevenuePoint:
      type: object
      required: [date, revenue, conversions]
      properties:
        date:
          type: string
        revenue:
          type: number
        conversions:
          type: integer
...