# lib/api-zod/src/generated/types/adminUserSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OfferLoots GPT Rewards Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUserSummary {
  id: number;
  username: string;
  email: string;
  status: string;
  role: string;
  balance: number;
  points: number;
  rank: string;
  totalEarned: number;
  totalConversions: number;
  createdAt: string;
}